Add authetication for add, update, delete products and removed for get products

+++ routes/products.py
@@ -12,7 +12,7 @@ from db_models.models import User
 router= APIRouter(
     prefix="/products",
     tags=["products"],
-    dependencies=[Depends(get_db), Depends(get_current_user)]
+    dependencies=[Depends(get_db)]
 )
 
 @router.get("/")
@@ -27,19 +27,21 @@ def get_product(product_id:int, session:Session = Depends(get_db)):
     return JSONResponse(content= {"data":data},status_code=200)
 
 @router.post('/')
-def add_product(new_product:ProductInput, session:Session = Depends(get_db)):
-    data = product_service.addProduct(session,new_product)
-    return JSONResponse(content={"msg":"Item added", "data":data},status_code=200)
+def add_product(new_product:ProductInput, session:Session = Depends(get_db),user:User =  Depends(get_current_user)):
+    if user is not None:
+        data = product_service.addProduct(session,new_product)
+        return JSONResponse(content={"msg":"Item added", "data":data},status_code=200)
 
 
 @router.put("/{product_id}")
-def update_product(product_id:int,product_up:ProductUpdate, session:Session = Depends(get_db)):
-    data = product_service.updateProduct(session,product_up,product_id=product_id)
-    return JSONResponse(content={"message":"item updated","data":data},status_code=200)
+def update_product(product_id:int,product_up:ProductUpdate, session:Session = Depends(get_db), user:User =  Depends(get_current_user)):
+    if user is not None:
+        data = product_service.updateProduct(session,product_up,product_id=product_id)
+        return JSONResponse(content={"message":"item updated","data":data},status_code=200)
 
 @router.delete('/{product_id}')
-def delete_product(product_id:int, session:Session = Depends(get_db)):
-    
-    item_id = product_service.deleteProduct(session,product_id)
-    return JSONResponse(content=f"Item with ID{item_id} deleted",status_code=200)
+def delete_product(product_id:int, session:Session = Depends(get_db), user:User =  Depends(get_current_user)):
+    if user is not None:
+        item_id = product_service.deleteProduct(session,product_id)
+        return JSONResponse(content=f"Item with ID{item_id} deleted",status_code=200)
